create classes for data and utility class for functions so that DataServlet is less cluttered
